+++ lizard_wms/models.py
@@ -31,6 +31,7 @@ RENDER_NONE = ''
 RENDER_TEXT = 'T'
 RENDER_IMAGE = 'I'
 RENDER_URL = 'U'
+RENDER_URL_MORE_LINK = 'M'
 RENDER_URL_LIKE = 'W'
 RENDER_GC_COLUMN = 'C'
 RENDER_XLS_DATE = 'X'
@@ -522,46 +523,14 @@ like {"key": "value", "key2": "value2"}.
     def get_popup_info(self, values):
         if not values:
             return
-
         info = []
-
         for feature_line in (self.featureline_set.filter(visible=True).
                              order_by('order_using')):
             if feature_line.name in values:
-                if feature_line.render_as == RENDER_GC_COLUMN:
-                    data = json.loads(values[feature_line.name])
-                    if data is None:
-                        # See https://github.com/nens/deltaportaal/issues/4
-                        logger.warn(
-                            "https://github.com/nens/deltaportaal/issues/4 "
-                            "hits again")
-                        return
-                    url = google_column_chart_url(data)
-                    values[feature_line.name] = url
-                    if url == '':
-                        feature_line.render_as = RENDER_NONE
-                    else:
-                        feature_line.render_as = RENDER_IMAGE
-                    feature_line.show_label = 'false'
-                elif feature_line.render_as == RENDER_XLS_DATE:
-                    data = json.loads(values[feature_line.name])
-                    if data is None:
-                        # See https://github.com/nens/deltaportaal/issues/4
-                        logger.warn(
-                            "https://github.com/nens/deltaportaal/issues/4 "
-                            "hits again")
-                        return
-                    values[feature_line.name] = xls_date_to_string(data)
-                    feature_line.render_as = RENDER_TEXT
-                    feature_line.show_label = 'true'
-                else:
-                    feature_line.show_label = 'true'
-                info.append(
-                    {'name': (feature_line.description or feature_line.name),
-                     'value': values[feature_line.name],
-                     'render_as': feature_line.render_as,
-                     'show_label': feature_line.show_label,
-                     })
+                popup_info = feature_line.as_popup_info(
+                    values[feature_line.name])
+                if popup_info:
+                    info.append(popup_info)
         return info
 
     @property
@@ -613,6 +582,7 @@ class FeatureLine(models.Model):
             (RENDER_XLS_DATE, _("Excel date format")),
             (RENDER_URL, "URL"),
             (RENDER_URL_LIKE, "URL-achtige tekst"),
+            (RENDER_URL_MORE_LINK, _("URL shown as 'click here' link")),
             (RENDER_GC_COLUMN, "Google column chart")), default=RENDER_TEXT)
     in_hover = models.BooleanField(default=False)
     order_using = models.IntegerField(default=1000)
@@ -630,6 +600,52 @@ class FeatureLine(models.Model):
         This gives us the most user-friendly name possible.
         """
         return self.description or self.name
+
+    def as_popup_info(self, value):
+        """Return ourselves as dict for in WMSSource's popup."""
+        show_label = 'true'
+        link_name = ''
+        if self.render_as == RENDER_GC_COLUMN:
+            json_data = json.loads(value)
+            if json_data is None:
+                # See https://github.com/nens/deltaportaal/issues/4
+                logger.warn(
+                    "https://github.com/nens/deltaportaal/issues/4 "
+                    "hits again")
+                return
+            url = google_column_chart_url(json_data)
+            value = url
+            if url == '':
+                self.render_as = RENDER_NONE
+            else:
+                self.render_as = RENDER_IMAGE
+            show_label = 'false'
+        elif self.render_as == RENDER_XLS_DATE:
+            try:
+                date_value = float(value)
+            except ValueError:
+                logger.warn("Not a float-like value for XLS date: %r", value)
+                return
+            value = xls_date_to_string(date_value)
+            self.render_as = RENDER_TEXT
+        elif self.render_as == RENDER_URL_LIKE:
+            link_name = value
+            value = 'http://' + value
+            # Quite brittle, but equal to the code from the template that it
+            # replaces.
+        elif self.render_as == RENDER_URL_MORE_LINK:
+            link_name = _("Click here for more information")
+            if not 'http://' in value:
+                value = 'http://' + value
+                # Yes, this means we could do the same for RENDER_URL_LIKE
+                # options, but I'm leaving that one alone for the moment.
+        return {
+            'name': (self.description or self.name),
+            'value': value,
+            'link_name': link_name,
+            'render_as': self.render_as,
+            'show_label': show_label,
+            }
 
 
 class FilterPage(models.Model):